app/requirements.txt
streamlit>=1.33
openai-whisper>=20231117
numpy>=1.24
tqdm>=4.66
pillow>=10.0
requests>=2.27
qrcode>=7.4
